requirements.txt
mysql-connector-python==8.0.23
protobuf==3.15.6
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
six==1.15.0
